Added some more turret logic

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utilities/MathUtils.java
@@ -2,7 +2,7 @@ package org.firstinspires.ftc.teamcode.utilities;
 
 public class MathUtils {
     // Clamp function for integers
-    public static int clamp(int value, int min, int max) {
+    public static double clamp( double value, double min, double max) {
         if (min > max) {
             throw new IllegalArgumentException("Min value must be less than or equal to max value");
         }
